// ts-apicalypse/test-data/compiled.d.ts
export namespace proto {

    interface ICount {
        count?: (number|null);
    }

    class Count implements ICount {
        constructor(properties?: proto.ICount);
        public count: number;
    }

    interface IMultiQueryResult {
        name?: (string|null);
        results?: (Uint8Array[]|null);
        count?: (number|null);
    }
// ...
    interface IAgeRating {
        id?: (number|null);
        category?: (proto.AgeRatingCategoryEnum|null);
        content_descriptions?: (proto.IAgeRatingContentDescription[]|null);
        rating?: (proto.AgeRatingRatingEnum|null);
        rating_cover_url?: (string|null);
        synopsis?: (string|null);
        checksum?: (string|null);
    }
// ...
}